requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
sqlalchemy>=2.0.0
alembic>=1.13.0
jinja2>=3.1.0
python-multipart>=0.0.9
aiofiles>=24.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
ldap3>=2.9.0
faster-whisper>=1.0.0
